Roadmap-genrator-main/requirements.txt
streamlit
google-generativeai
python-dotenv
pillow
pandas==2.2.0
scikit-learn==1.4.0
numpy==1.26.0
scikit-learn